fix(azure): cd ls command used wrong blob container
